Use Int32 for tag, source, dest

+++ ext/ReactantMPIExt/Overrides.jl
@@ -54,8 +54,8 @@ function MPI.Isend(
     comm::MPI.Comm,
     request::TracedRequest=TracedRequest((), nothing),
 )
-    dest = Reactant.Ops.constant(dest)
-    tag = Reactant.Ops.constant(tag)
+    dest = Reactant.Ops.constant(Int32(dest))
+    tag = Reactant.Ops.constant(Int32(tag))
 
     gen_request = MPI.Isend(buf, dest, tag, comm)
     request.mlir_data = gen_request.mlir_data
@@ -72,8 +72,8 @@ function MPI.Isend(
 end
 
 function MPI.Recv!(buf::TracedRArray, source::Integer, tag::Integer, comm::MPI.Comm)
-    tag = Reactant.Ops.constant(tag)
-    source = Reactant.Ops.constant(source)
+    tag = Reactant.Ops.constant(Int32(tag))
+    source = Reactant.Ops.constant(Int32(source))
     return MPI.Recv!(buf, source, tag, comm)
 end
 
@@ -109,8 +109,8 @@ function MPI.Irecv!(
     comm::MPI.Comm,
     request::TracedRequest=TracedRequest((), nothing),
 )
-    source = Reactant.Ops.constant(source)
-    tag = Reactant.Ops.constant(tag)
+    source = Reactant.Ops.constant(Int32(source))
+    tag = Reactant.Ops.constant(Int32(tag))
 
     gen_request = MPI.Irecv!(buf, source, tag, comm)
     request.mlir_data = gen_request.mlir_data
